fix broadcastchannel polyfill in service worker
The problem at hand was since service workers persist even during page loads, that means the service worker would not have access
to the polyfilled `BroadcastChannel`, and as a result, it wouldn't work on browsers such as Safari. To fix this, I have included the
polyfill in the service worker bundle. The two downsides to this is 1) a slightly larger service worker file by around ~1BKB, and that
now the polyfill is loaded twice.