use esbuild's `inject` option instead of importing polyfill
The problem when using `import` in the service worker to import the polyfill was that using esbuild's `bundle` option was required to be used.
Now this was a problem because esbuild requires we either build our service worker as either an IIFE, a CommonJS module, or as an ESM. The
`inject` option circumvents this problem.